Revert "vice 3.10"
This reverts commit af4f64528798a13301728f0e291cbb8d7eaff617.

+++ src/arch/sdl/menu_common.c
@@ -245,8 +245,8 @@ char win32_path_buf[80];
 
 static char *sdl_ui_menu_file_translate_seperator(const char *text)
 {
-    size_t len;
-    size_t i;
+    int len;
+    int i;
 
     len = strlen(text);
